requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-modeltranslation==0.15
django-recaptcha3==0.2.0
flower==0.9.4
humanize==0.5.1
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
Pillow==7.1.1
pkg-resources==0.0.0
python-decouple==3.3
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
